+++ mail/mail-common/src/models/conversation.rs
@@ -35,8 +35,11 @@ use proton_api_mail::services::proton::response_data::{
     Conversation as ApiConversation, ConversationLabel as ApiConversationLabel,
     MessageMetadata as ApiMessageMetadata, OperationResult,
 };
-use proton_core_common::datatypes::{LabelType, LocalLabelId, SystemLabel};
-use proton_core_common::models::{Label, ModelExtension, ModelIdExtension};
+use proton_core_common::datatypes::{InitializationKey, LabelType, LocalLabelId, SystemLabel};
+use proton_core_common::models::{
+    InitializationError, InitializationWatcher, InitializedComponent, Label, ModelExtension,
+    ModelIdExtension,
+};
 use proton_core_common::utils::MapVec as _;
 use proton_mail_ids::LocalConversationId;
 use sqlite_watcher::watcher::TableObserver;
@@ -50,6 +53,7 @@ use std::collections::hash_map::Entry as HmEntry;
 use std::collections::{BTreeSet, HashMap, HashSet};
 use std::future::Future;
 use std::ops::AddAssign;
+use std::sync::Arc;
 use tracing::{debug, error, info, warn};
 
 #[derive(Clone, Debug, Default, Eq, Model, PartialEq)]
@@ -3558,26 +3562,42 @@ impl ConversationCounters {
     }
 }
 
-// TODO: Refactor this at 1.85 into an AsyncFnOnce
-/// This acts like a closure that on `new` syncs the label counters and on `store` actually stores
-/// them to the db.
-/// This is useful to split the store and fetch behavior since we need labels to exist in the db
-/// before storing them.
+/// Used to initialize counters by syncing it with the Backend
 pub struct StoreLabelCounters(Vec<ConversationLabelsCount>, Vec<MessageLabelsCount>);
 impl StoreLabelCounters {
-    pub async fn new(api: &impl ProtonMail) -> Result<Self, AppError> {
-        let (a, b) =
-            future::try_join(Conversation::fetch_counts(api), Message::fetch_counts(api)).await?;
-        Ok(Self(a, b))
-    }
+    /// Key used to distinguish between components in the initialization.
+    /// It is a string, not an enum for making it open for additional changes from different BU.
+    ///
+    pub const INIT_KEY: InitializationKey = InitializationKey::new("label_counters");
 
-    pub async fn store(self, stash: &Stash) -> Result<(), AppError> {
-        let mut tether = stash.connection();
-        let tx = tether.transaction().await?;
-        ConversationLabelsCount::create_or_update_conversation_counts(self.0, &tx).await?;
-        MessageLabelsCount::create_or_update_message_counts(self.1, &tx).await?;
-        tx.commit().await?;
-        Ok(())
+    /// It initializes counters by syncing with the Backend.
+    /// In case of successful initialization, it marks it in the [`InitializedComponents`].
+    ///
+    /// This function is idempotent. If successfully initialized in the past.
+    ///
+    pub async fn initialize(
+        watcher: Arc<InitializationWatcher>,
+        api: &impl ProtonMail,
+        stash: &Stash,
+    ) -> Result<(), InitializationError<AppError>> {
+        InitializedComponent::initialize::<AppError, Self>(
+            watcher,
+            Self::INIT_KEY,
+            &[Label::INIT_KEY],
+            stash.connection(),
+            async || {
+                let (a, b) =
+                    future::try_join(Conversation::fetch_counts(api), Message::fetch_counts(api))
+                        .await?;
+                Ok(Self(a, b))
+            },
+            async |tx, Self(a, b)| {
+                ConversationLabelsCount::create_or_update_conversation_counts(a, tx).await?;
+                MessageLabelsCount::create_or_update_message_counts(b, tx).await?;
+                Ok(())
+            },
+        )
+        .await
     }
 }